core-engine/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
boto3>=1.34.0
httpx>=0.25.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
